src/main/java/com/petSmile/DatabaseConnection.java
package com.petSmile;
import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

public class DatabaseConnection {
    private static final String URL = "jdbc:mariadb://localhost:3306/clinica?useUnicode=true&characterEncoding=UTF-8";
    private static final String USER = "root"; 
    private static final String PASSWORD = "admin";

    public static Connection getConnection() throws SQLException {
        return DriverManager.getConnection(URL, USER, PASSWORD);
    }
}
